refactor, remove bootstrap from makefile
