+++ cli/command/cluster/cluster.go
@@ -88,3 +88,27 @@ func stripPrefixes(cmd *cobra.Command, args []string) []string {
 	})
 	return args
 }
+
+func runPluginCommand(c cli.Interface, cmd *cobra.Command, command string) error {
+	// args and env will be supplied to the cluster plugin container
+	var args []string
+	var env map[string]string
+	//	env = map[string]string{"TAG": opts.tag, "REGISTRATION": opts.registration, "NOTIFICATIONS": strconv.FormatBool(opts.notifications)}
+
+	args = append(args, command)
+	args = stripPrefixes(cmd, args)
+
+	config := PluginConfig{
+		Provider:   opts.provider,
+		DockerOpts: opts.docker,
+		// TODO: not clear yet if we'll need this
+		Options: opts.options,
+	}
+
+	p, err := NewPlugin(config)
+	if err != nil {
+		return err
+	}
+
+	return p.Run(c, args, env)
+}
